post rebase cleanup
