+++ rust/feature-flags/src/client/mod.rs
@@ -1,3 +1,1 @@
 pub mod database;
-
-pub mod redis;
