+++ rust/feature-flags/src/router.rs
@@ -18,8 +18,7 @@ use tower_http::{
 
 use crate::{
     api::{endpoint, test_endpoint},
-
-    client::{database::Client as DatabaseClient, redis::Client as RedisClient},
+    client::database::Client as DatabaseClient,
     cohort::cohort_cache_manager::CohortCacheManager,
     config::{Config, TeamIdsToTrack},
     metrics::metrics_utils::team_id_label_filter,
